refactor: `Workspace` -> `DocCollection`
